Add: Gameboard file and tests
When this Gameboard object places a ship, it creates a Ship object
and places the same object in several cells, based on its length.

The grid size is 10x10, but can be changed later to use another custom
size.

In order to make the receiveAttack tests, the function signature
contains a callback parameter. It calls this function when it detects
a missing attack.

The missedAttacks variable also is public to prevent errors in the test.
At the moment of writing this, I don't know how to make it private and
use a function instead.