Feature: make game playable
Added dom and game modules. There are some changes to the previous
factory objects to allow certain parts of the game loop. In ship.js,
removed sunk property since it's no longer used.

In player.js, changed the way a computer player calculates a new attack.
Now it has a attacksMade array that gets updated when it detects a pair
of coordinates that hasn't calculated before.

Also changed the player.test.js related to calculateAttack function to
keep them working.

In gameboard.js, the receiveAttack function now returns true or false
based on the result. This is done to render the cells.

Added some style to let the user play. There are things to be improved,
like more style to the rest of the page and a better style when the
user loses a game.