Change: calculateAttack function in player.js
This function now gets an adjacent cell coordinates after a successful
attack. The previous loop to get a random pair of coordinates is still
used and made another function to be reusable in this same function.

The attacksMade array now contains an object that contains the
coordinates object used previously and the result property that's used
to know if the previous attack was hit or not.

A new registerAttack function is added and made public to be used in
the computerTurn function in dom.js where the hit's result is
calculated. This function just adds an object to attacksMade array.

There are other new functions in this file, all of them to ease this new
behavior to get adjacent cells after a hit.