Feature: let player place his ships
Added a form to let the player place a ship. It has some restrictions,
like a ship cannot be placed next to another one or to collide with
other. Also, it's coded to let a max ship count (5 at the moment), to
ensure that both players have the same amount of ships. After this, the
form is deleted and the enemy gameboard is prepared to start the game.

Some of the functions needed to do this are in the dom.js, like
everything related to dom stuff and the majority of logic things in the
gameboard.js.

Removed the populateGameboard call for human gamboard in game.js. It's
no needed anymore.

At the moment of writing this, the computer gameboard still has the
same ships every time. It will be added functionality to randomly
place its ships.